+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added `tensors.FromScalar`
 * Updated to gopjrt v0.3.0
 * Package `graph`:
+  * Added `ExecOnce` and `ExecOnceN` 
   * Added `CumSum`
 
 ## 0.11.0 BREAKING CHANGE: Multi-Backend support; Added XLA/PJRT support (with gopjrt); meaningful speed ups; No more C code (all goes through gopjrt)